LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/plaintext-only/insertLineBreak_white-space=normal-expected.txt

AB


PASS execCommand("insertLineBreak") when A[]B
PASS Pressing Enter when A[]B should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when A[]B should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <p>{}<br></p>
PASS Pressing Enter when <p>{}<br></p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <p>{}<br></p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <p style="white-space:normal">A[]B</p>
PASS Pressing Enter when <p style="white-space:normal">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <p style="white-space:normal">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <p style="white-space:pre">A[]B</p>
PASS Pressing Enter when <p style="white-space:pre">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <p style="white-space:pre">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <p style="white-space:pre-line">A[]B</p>
PASS Pressing Enter when <p style="white-space:pre-line">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <p style="white-space:pre-line">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <p style="white-space:pre-wrap">A[]B</p>
PASS Pressing Enter when <p style="white-space:pre-wrap">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <p style="white-space:pre-wrap">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <ul><li>[]AB</li></ul>
PASS Pressing Enter when <ul><li>[]AB</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <ul><li>[]AB</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <ul><li>A[]B</li></ul>
PASS Pressing Enter when <ul><li>A[]B</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <ul><li>A[]B</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <ul><li>AB[]</li></ul>
PASS Pressing Enter when <ul><li>AB[]</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <ul><li>AB[]</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <dl><dt>[]AB</dt></dl>
PASS Pressing Enter when <dl><dt>[]AB</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <dl><dt>[]AB</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <dl><dt>A[]B</dt></dl>
PASS Pressing Enter when <dl><dt>A[]B</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <dl><dt>A[]B</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <dl><dt>AB[]</dt></dl>
PASS Pressing Enter when <dl><dt>AB[]</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <dl><dt>AB[]</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <dl><dd>[]AB</dd></dl>
PASS Pressing Enter when <dl><dd>[]AB</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <dl><dd>[]AB</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <dl><dd>A[]B</dd></dl>
PASS Pressing Enter when <dl><dd>A[]B</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <dl><dd>A[]B</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <dl><dd>AB[]</dd></dl>
PASS Pressing Enter when <dl><dd>AB[]</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <dl><dd>AB[]</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <table><tbody><tr><td>[]AB</td></tr></tbody></table>
PASS Pressing Enter when <table><tbody><tr><td>[]AB</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <table><tbody><tr><td>[]AB</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <table><tbody><tr><td>A[]B</td></tr></tbody></table>
PASS Pressing Enter when <table><tbody><tr><td>A[]B</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <table><tbody><tr><td>A[]B</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <table><tbody><tr><td>AB[]</td></tr></tbody></table>
PASS Pressing Enter when <table><tbody><tr><td>AB[]</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <table><tbody><tr><td>AB[]</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <h1>[]AB</h1>
PASS Pressing Enter when <h1>[]AB</h1> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <h1>[]AB</h1> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <h1>A[]B</h1>
PASS Pressing Enter when <h1>A[]B</h1> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <h1>A[]B</h1> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS execCommand("insertLineBreak") when <h1>AB[]</h1>
PASS Pressing Enter when <h1>AB[]</h1> should cause "insertLineBreak" and shouldn't insert new paragraph
PASS Pressing Shift+Enter when <h1>AB[]</h1> should cause "insertLineBreak" and shouldn't insert new paragraph
FAIL Pressing Enter when A[]B should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when A[]B should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "A<br>B" but got "AB"
FAIL Pressing Shift+Enter when A[]B should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when A[]B should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "A<br>B" but got "AB"
FAIL Pressing Enter when <p>{}<br></p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <p>{}<br></p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<p><br><br></p>" but got "<p><br></p>"
FAIL Pressing Shift+Enter when <p>{}<br></p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <p>{}<br></p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<p><br><br></p>" but got "<p><br></p>"
FAIL Pressing Enter when <p style="white-space:normal">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <p style="white-space:normal">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<p style=\"white-space:normal\">A<br>B</p>" but got "<p style=\"white-space:normal\">AB</p>"
FAIL Pressing Shift+Enter when <p style="white-space:normal">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <p style="white-space:normal">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<p style=\"white-space:normal\">A<br>B</p>" but got "<p style=\"white-space:normal\">AB</p>"
FAIL Pressing Enter when <p style="white-space:pre">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <p style="white-space:pre">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_in_array: value "<p style=\"white-space:pre\">AB</p>" not in array ["<p style=\"white-space:pre\">A<br>B</p>", "<p style=\"white-space:pre\">A\nB</p>"]
FAIL Pressing Shift+Enter when <p style="white-space:pre">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <p style="white-space:pre">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_in_array: value "<p style=\"white-space:pre\">AB</p>" not in array ["<p style=\"white-space:pre\">A<br>B</p>", "<p style=\"white-space:pre\">A\nB</p>"]
FAIL Pressing Enter when <p style="white-space:pre-line">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <p style="white-space:pre-line">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_in_array: value "<p style=\"white-space:pre-line\">AB</p>" not in array ["<p style=\"white-space:pre-line\">A<br>B</p>", "<p style=\"white-space:pre-line\">A\nB</p>"]
FAIL Pressing Shift+Enter when <p style="white-space:pre-line">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <p style="white-space:pre-line">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_in_array: value "<p style=\"white-space:pre-line\">AB</p>" not in array ["<p style=\"white-space:pre-line\">A<br>B</p>", "<p style=\"white-space:pre-line\">A\nB</p>"]
FAIL Pressing Enter when <p style="white-space:pre-wrap">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <p style="white-space:pre-wrap">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_in_array: value "<p style=\"white-space:pre-wrap\">AB</p>" not in array ["<p style=\"white-space:pre-wrap\">A<br>B</p>", "<p style=\"white-space:pre-wrap\">A\nB</p>"]
FAIL Pressing Shift+Enter when <p style="white-space:pre-wrap">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <p style="white-space:pre-wrap">A[]B</p> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_in_array: value "<p style=\"white-space:pre-wrap\">AB</p>" not in array ["<p style=\"white-space:pre-wrap\">A<br>B</p>", "<p style=\"white-space:pre-wrap\">A\nB</p>"]
FAIL Pressing Enter when <ul><li>[]AB</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <ul><li>[]AB</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<ul><li><br>AB</li></ul>" but got "<ul><li>AB</li></ul>"
FAIL Pressing Shift+Enter when <ul><li>[]AB</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <ul><li>[]AB</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<ul><li><br>AB</li></ul>" but got "<ul><li>AB</li></ul>"
FAIL Pressing Enter when <ul><li>A[]B</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <ul><li>A[]B</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<ul><li>A<br>B</li></ul>" but got "<ul><li>AB</li></ul>"
FAIL Pressing Shift+Enter when <ul><li>A[]B</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <ul><li>A[]B</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<ul><li>A<br>B</li></ul>" but got "<ul><li>AB</li></ul>"
FAIL Pressing Enter when <ul><li>AB[]</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <ul><li>AB[]</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<ul><li>AB<br><br></li></ul>" but got "<ul><li>AB</li></ul>"
FAIL Pressing Shift+Enter when <ul><li>AB[]</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <ul><li>AB[]</li></ul> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<ul><li>AB<br><br></li></ul>" but got "<ul><li>AB</li></ul>"
FAIL Pressing Enter when <dl><dt>[]AB</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <dl><dt>[]AB</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dt><br>AB</dt></dl>" but got "<dl><dt>AB</dt></dl>"
FAIL Pressing Shift+Enter when <dl><dt>[]AB</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <dl><dt>[]AB</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dt><br>AB</dt></dl>" but got "<dl><dt>AB</dt></dl>"
FAIL Pressing Enter when <dl><dt>A[]B</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <dl><dt>A[]B</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dt>A<br>B</dt></dl>" but got "<dl><dt>AB</dt></dl>"
FAIL Pressing Shift+Enter when <dl><dt>A[]B</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <dl><dt>A[]B</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dt>A<br>B</dt></dl>" but got "<dl><dt>AB</dt></dl>"
FAIL Pressing Enter when <dl><dt>AB[]</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <dl><dt>AB[]</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dt>AB<br><br></dt></dl>" but got "<dl><dt>AB</dt></dl>"
FAIL Pressing Shift+Enter when <dl><dt>AB[]</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <dl><dt>AB[]</dt></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dt>AB<br><br></dt></dl>" but got "<dl><dt>AB</dt></dl>"
FAIL Pressing Enter when <dl><dd>[]AB</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <dl><dd>[]AB</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dd><br>AB</dd></dl>" but got "<dl><dd>AB</dd></dl>"
FAIL Pressing Shift+Enter when <dl><dd>[]AB</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <dl><dd>[]AB</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dd><br>AB</dd></dl>" but got "<dl><dd>AB</dd></dl>"
FAIL Pressing Enter when <dl><dd>A[]B</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <dl><dd>A[]B</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dd>A<br>B</dd></dl>" but got "<dl><dd>AB</dd></dl>"
FAIL Pressing Shift+Enter when <dl><dd>A[]B</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <dl><dd>A[]B</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dd>A<br>B</dd></dl>" but got "<dl><dd>AB</dd></dl>"
FAIL Pressing Enter when <dl><dd>AB[]</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <dl><dd>AB[]</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dd>AB<br><br></dd></dl>" but got "<dl><dd>AB</dd></dl>"
FAIL Pressing Shift+Enter when <dl><dd>AB[]</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <dl><dd>AB[]</dd></dl> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<dl><dd>AB<br><br></dd></dl>" but got "<dl><dd>AB</dd></dl>"
FAIL Pressing Enter when <table><tbody><tr><td>[]AB</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <table><tbody><tr><td>[]AB</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<table><tbody><tr><td><br>AB</td></tr></tbody></table>" but got "<table><tbody><tr><td>AB</td></tr></tbody></table>"
FAIL Pressing Shift+Enter when <table><tbody><tr><td>[]AB</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <table><tbody><tr><td>[]AB</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<table><tbody><tr><td><br>AB</td></tr></tbody></table>" but got "<table><tbody><tr><td>AB</td></tr></tbody></table>"
FAIL Pressing Enter when <table><tbody><tr><td>A[]B</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <table><tbody><tr><td>A[]B</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<table><tbody><tr><td>A<br>B</td></tr></tbody></table>" but got "<table><tbody><tr><td>AB</td></tr></tbody></table>"
FAIL Pressing Shift+Enter when <table><tbody><tr><td>A[]B</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <table><tbody><tr><td>A[]B</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<table><tbody><tr><td>A<br>B</td></tr></tbody></table>" but got "<table><tbody><tr><td>AB</td></tr></tbody></table>"
FAIL Pressing Enter when <table><tbody><tr><td>AB[]</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <table><tbody><tr><td>AB[]</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<table><tbody><tr><td>AB<br><br></td></tr></tbody></table>" but got "<table><tbody><tr><td>AB</td></tr></tbody></table>"
FAIL Pressing Shift+Enter when <table><tbody><tr><td>AB[]</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <table><tbody><tr><td>AB[]</td></tr></tbody></table> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<table><tbody><tr><td>AB<br><br></td></tr></tbody></table>" but got "<table><tbody><tr><td>AB</td></tr></tbody></table>"
FAIL Pressing Enter when <h1>[]AB</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <h1>[]AB</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<h1><br>AB</h1>" but got "<h1>AB</h1>"
FAIL Pressing Shift+Enter when <h1>[]AB</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <h1>[]AB</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<h1><br>AB</h1>" but got "<h1>AB</h1>"
FAIL Pressing Enter when <h1>A[]B</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <h1>A[]B</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<h1>A<br>B</h1>" but got "<h1>AB</h1>"
FAIL Pressing Shift+Enter when <h1>A[]B</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <h1>A[]B</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<h1>A<br>B</h1>" but got "<h1>AB</h1>"
FAIL Pressing Enter when <h1>AB[]</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Enter when <h1>AB[]</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<h1>AB<br><br></h1>" but got "<h1>AB</h1>"
FAIL Pressing Shift+Enter when <h1>AB[]</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: beforeinput assert_equals: inputType should be "insertLineBreak" expected (string) "insertLineBreak" but got (undefined) undefined
FAIL Pressing Shift+Enter when <h1>AB[]</h1> should cause "insertLineBreak" and shouldn't insert new paragraph: editingHost.innerHTML assert_equals: expected "<h1>AB<br><br></h1>" but got "<h1>AB</h1>"
